+++ src/react-server-dom-webpack/cjs/react-server-dom-webpack-server.browser.production.js
@@ -1137,7 +1137,8 @@ function renderElement(request, task, type, key, ref, props) {
       case REACT_MEMO_TYPE:
         return renderElement(request, task, type.type, key, ref, props);
     }
-  type.$$typeof === CLIENT_REFERENCE_TAG$1 &&
+  null != type &&
+    type.$$typeof === CLIENT_REFERENCE_TAG$1 &&
     emitClientReferenceCSS(request, type);
   request = key;
   key = task.keyPath;